feat(notifications)!: gravar o corpo do callback uma vez
A ingestão replicava o envelope do callback em cada linha de evento do lote, o
que tornava escrita e leitura quadráticas num tamanho que quem escolhe é o
provedor. A tabela delivery_payload passa a guardar o corpo uma vez,
particionada por mês em received_at nos mesmos limites da evidência, e cada
delivery_event a referencia por payload_id.

A propriedade de que a evidência de um evento é o lote inteiro passa a ser
garantida pela ordem de escrita e não por transação compartilhada: a linha de
payload confirma no máximo junto do primeiro evento que a referencia, o que
mantém uma transação por evento e deixa um lote inteiramente reentregue sem
gravar nada. O instante da recepção passa a ser carimbado uma vez por callback,
que é o que mantém o lote dentro de uma única partição mensal nas duas tabelas.

O teto de bytes deixa de existir em duplicata: as options de ingestão passam a
ser as donas do valor. O teto de eventos permanece, agora justificado pelo que
de fato limita, que é o número de transações de uma requisição, e não mais pelo
volume gravado.

A sonda mede o resultado: a amplificação de escrita fica plana entre 3,6x e
4,6x de lotes de 1 a 500 eventos, quando antes acompanharia o tamanho do lote.

BREAKING CHANGE: delivery_event perde a coluna payload_enc e ganha payload_id
NOT NULL. Não há backfill: uma tabela que já guarde evidência recusa a migração
em vez de inventar um callback para linhas cujos bytes não sabe atribuir.

+++ src/Platform.Api/Modules/Notifications/Infrastructure/Authentication/ProviderSignatureAuthenticationHandler.Logger.cs
@@ -34,7 +34,7 @@ internal static partial class ProviderSignatureAuthenticationHandlerLogger
         Level = LogLevel.Warning,
         Message = "Callback de '{ProviderKey}' recusado por exceder o limite de {MaxBodyBytes} bytes.")]
     internal static partial void ProviderWebhookBodyTooLarge(
-        this ILogger logger, string providerKey, int maxBodyBytes);
+        this ILogger logger, string providerKey, long maxBodyBytes);
 
     [LoggerMessage(
         EventId = 7184,
